Allow live views to self-update
The `CoroutineScope.actor()` API from `kotlinx.coroutines.cannels` might
be a better fit even though it is marked as `@ObsoleteCoroutinesApi`.

Kotlin might get a new actors API at some point. It is unclear though,
when that might happen. Maybe it is worth rolling with the current one
and porting onto the new API when and if it comes out.

https://github.com/Kotlin/kotlinx.coroutines/issues/87